feat(cpu widget): allow user to switch (default) processor core view
Using the new CPU_CORES_TOGGLE_MODE env option the user is now able to specify which Processor
core view it want to see. The available options are 'toggle', 'multi-core' and 'average'. The
'toggle' option is default and has the same functionality/effect as before. The 'multi-core' option
and 'average' option switches to the corresponding view while hiding the toggle.